Refactoring Reversi

+++ Reversi/src/main/java/player/ai/AlphaBetaEvaluationSimpleAI.java
@@ -10,24 +10,24 @@ import board.Turn;
 
 public class AlphaBetaEvaluationSimpleAI extends Player {
     private final int maxDepth;
-    
+
     public AlphaBetaEvaluationSimpleAI(Turn turn, int maxDepth) {
         super(turn);
         this.maxDepth = maxDepth;
     }
-    
+
     @Override
     public Position play(Board board) {
         return eval(board, maxDepth, turn.stone(), new BoardScoreEvaluation(turn), Integer.MIN_VALUE, Integer.MAX_VALUE).getPosition();
     }
-    
-    protected EvalResult eval(Board board, int restDepth, Stone stone, Evaluation evaluation, int alpha, int beta) {
+
+    protected EvalResult eval(Board board, int restDepth, Stone stone, Evaluation evaluation, double alpha, double beta) {
         if (restDepth == 0)
             return new EvalResult(evaluation.score(board, stone), null);
-        
+
         // 自分の番では、評価が最も大きくなるものを選ぶ。
         boolean didPlayed = false;
-        int maxScore = -10000;
+        double maxScore = -10000;
         Position p = null;
         for (int y = 1; y <= Board.HEIGHT; ++y) {
             for (int x = 1; x <= Board.WIDTH; ++x) {
@@ -36,25 +36,25 @@ public class AlphaBetaEvaluationSimpleAI extends Player {
                     Board b = new Board(board);
                     Evaluation e = evaluation.clone();
                     e.willPut(board, x, y, stone);
-                    b.put(x, y, stone);                    
-                    int score = -eval(b, restDepth - 1, stone.flip(), e, -beta, -Math.max(alpha, maxScore)).getScore();
+                    b.put(x, y, stone);
+                    double score = -eval(b, restDepth - 1, stone.flip(), e, -beta, -Math.max(alpha, maxScore)).getScore();
                     if (maxScore < score) {
                         maxScore = score;
                         p = new Position(x, y);
                     }
-                    
+
                     // beta cut
                     if (maxScore >= beta)
                         return new EvalResult(score, p);
                 }
             }
         }
-        
+
         if (didPlayed) {
             assert (p != null);
             return new EvalResult(maxScore, p);
         } else {
-            int score = -eval(board, restDepth - 1, stone.flip(), evaluation, -beta, -alpha).getScore();
+            double score = -eval(board, restDepth - 1, stone.flip(), evaluation, -beta, -alpha).getScore();
             return new EvalResult(score, null);
         }
     }
