+++ Reversi/src/main/java/player/ai/NegaMaxSimpleAI.java
@@ -17,14 +17,14 @@ public class NegaMaxSimpleAI extends Player {
         { -50, -70, 10, 15, 15, 10, -70, -50 },
         { 100, -50, 35, 30, 30, 35, -50, 100 },
     };
-    
+
     private final int maxDepth;
-    
+
     public NegaMaxSimpleAI(Turn turn, int maxDepth) {
         super(turn);
         this.maxDepth = maxDepth;
     }
-    
+
     @Override
     public Position play(Board board) {
         return eval(board, maxDepth, turn.stone(), 0).getPosition();
@@ -33,10 +33,10 @@ public class NegaMaxSimpleAI extends Player {
     private EvalResult eval(Board board, int restDepth, Stone stone, int scoreSum) {
         if (restDepth == 0)
             return new EvalResult(-scoreSum, null);
-        
+
         // 自分の番では、評価が最も大きくなるものを選ぶ。
         boolean didPlayed = false;
-        int maxScore = -10000;
+        double maxScore = -10000;
         Position p = null;
         for (int y = 1; y <= Board.HEIGHT; ++y) {
             for (int x = 1; x <= Board.WIDTH; ++x) {
@@ -44,7 +44,7 @@ public class NegaMaxSimpleAI extends Player {
                     didPlayed = true;
                     Board b = new Board(board);
                     b.put(x, y, stone);
-                    int score = -eval(b, restDepth - 1, stone.flip(), scoreSum + (stone == turn.stone() ? EVAL_VALUES[y-1][x-1] : -EVAL_VALUES[y-1][x-1])).getScore();
+                    double score = -eval(b, restDepth - 1, stone.flip(), scoreSum + (stone == turn.stone() ? EVAL_VALUES[y-1][x-1] : -EVAL_VALUES[y-1][x-1])).getScore();
                     if (maxScore < score) {
                         maxScore = score;
                         p = new Position(x, y);
@@ -52,13 +52,13 @@ public class NegaMaxSimpleAI extends Player {
                 }
             }
         }
-        
+
         if (didPlayed) {
             assert (p != null);
             return new EvalResult(maxScore, p);
         } else {
             // 自分の番がパスだった場合は、-100 点しておく。
-            int score = -eval(board, restDepth - 1, stone.flip(), scoreSum).getScore();
+            double score = -eval(board, restDepth - 1, stone.flip(), scoreSum).getScore();
             return new EvalResult(score, null);
         }
     }
